New drawing method
I added a new means of drawing using Shapes.
Shapes are a single-line so far, each one containing a number of points
in said line.
Each point in the line (ShapeBlock) has whether or not it should draw in
it. Each shape does the same thing. When the user input finishes, the
shape is considered completed, and will not render until playback.